+++ src/MoptWorldline.php
@@ -24,7 +24,7 @@ use Shopware\Core\Framework\Plugin\Util\PluginIdProvider;
 class MoptWorldline extends Plugin
 {
     const PLUGIN_NAME = 'MoptWorldline';
-    const PLUGIN_VERSION = '3.2.7';
+    const PLUGIN_VERSION = '3.2.9';
     const PLUGIN_ID = 'MoptWorldline';
     const PLUGIN_CREATOR = 'Mediaopt GmbH';
     private const ENCRYPTION_INTRODUCED_VERSION = '3.2.7';